refactor: isolate runtime command dispatch
